add accordion, add data to validatie yup, remove accordion, fix subcategories in thridstep
