+++ src/components/products/ProductDetails.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const ProductDetails = () => {
+    return (
+        <div>ProductDetails</div>
+    )
+}
