+++ src/components/products/ProductView.tsx
@@ -119,75 +119,6 @@ const ProductView: FC<Product | any> = ({
           <Text fontWeight={500} fontSize='md' mt={4}>
             Garantía de 30 días
           </Text>
-          <Accordion marginTop={'10px'} allowToggle >
-            <AccordionItem>
-              <h2>
-                <AccordionButton>
-                  <Box fontWeight={700} as="span" flex='1' textAlign='left'>
-                    Caracteristicas adicionales
-                  </Box>
-                  <AccordionIcon />
-                </AccordionButton>
-              </h2>
-              <AccordionPanel>
-                {Object.keys(productt ?? {}).map((element, index) => {
-                  const unidad = productt[`unidad-${element}`]
-                  const contentProduct = productt[element]
-                  const sectionsExclude = ['category', 'description', 'price', 'tags', 'stock', 'title', 'image', 'subCategory']
-
-                  const returnDefault = (title: string, content: string) => {
-                    return (
-                      <Box display={'flex'} sx={{ alignItems: 'flex-end' }} key={index}>
-                        <Text fontWeight={700} fontSize='lg' mt={4}>{title + ": "}</Text>
-                        <Text fontWeight={500}>{content}</Text>
-                      </Box>
-                    )
-                  }
-
-                  if (element === 'Graduacion') {
-                    const longitudGrupo = Math.ceil((contentProduct?.length ?? 3) / 3);
-                    const newData = [contentProduct.substring(0, longitudGrupo), contentProduct.substring(longitudGrupo, longitudGrupo * 2), contentProduct.substring(longitudGrupo * 2)]
-                    return (
-                      <Box>
-                        <Text fontWeight={700} fontSize='lg' mt={4}>{element}</Text>
-                        {newData.map((e) => (
-                          <Badge sx={{ marginRight: '20px' }} variant='subtle' colorScheme='green' key={index * 3} >
-                            {e}
-                          </Badge>
-                        ))
-                        }
-                      </Box>
-                    )
-
-                  }
-                  if (element.includes('unidad') || contentProduct?.length === 0 || element === 'id' || sectionsExclude.includes(element)) return null
-                  if (Array.isArray(contentProduct)) {
-                    return (
-                      <Box>
-                        <Text fontWeight={700} fontSize='lg' mt={4}>{element}</Text>
-                        {
-                          contentProduct.map((e: [], i: any) => (
-                            <Badge sx={{ marginRight: '20px' }} variant='outline' colorScheme='green' key={i} >
-                              {e}
-                            </Badge>
-                          ))
-                        }
-                      </Box>
-                    )
-                  }
-                  if (productt[`unidad-${element}`] !== undefined) {
-                    return returnDefault(element, (contentProduct + " " + unidad))
-                  }
-                  return (
-                    <Box key={index + 1}>
-                      {returnDefault(element, contentProduct)}
-                    </Box>
-                  )
-                })
-                }
-              </AccordionPanel>
-            </AccordionItem>
-          </Accordion>
         </Box>
         <Stack
           direction={["row", "row", "column"]}
