+++ src/pages/products/Product.tsx
@@ -1,31 +1,10 @@
 import { FC, lazy, Suspense, useContext, useEffect } from "react";
-
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  BreadcrumbLink,
-  Center,
-  Box,
-  VStack,
-  Heading,
-  Stack,
-  Text,
-  Button,
-  useToast,
-  useMediaQuery,
-  IconButton,
-  Input,
-  FormControl,
-  FormLabel,
-  Textarea,
-  FormErrorMessage,
-} from "@chakra-ui/react";
+import { Breadcrumb, BreadcrumbItem, BreadcrumbLink, Center, Box, VStack, Heading, Stack, Text, Button, useToast, useMediaQuery, IconButton, Input, FormControl, FormLabel, Textarea, FormErrorMessage, Accordion, AccordionItem, AccordionPanel, AccordionButton, AccordionIcon, Badge, } from "@chakra-ui/react";
 import { Link, useNavigate, useParams } from "react-router-dom";
 import { FiArrowLeft } from "react-icons/fi";
 import { Helmet } from "react-helmet-async";
 import { SubmitHandler, useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
-import * as yup from "yup";
 
 import { ProductSkeleton } from "@/components/skeleton";
 import { Comment } from "@/components/ui";
@@ -33,6 +12,8 @@ import { useComments, useProduct } from "@/hooks";
 import { CommentInfo } from "@/interfaces";
 import { addComment } from "@/utils";
 import { UserContext } from "@/context";
+import { AddIcon } from "@chakra-ui/icons";
+import { userSchema, guestSchema } from './product.data';
 
 const ProductView = lazy(() => import("@/components/products/ProductView"));
 
@@ -49,29 +30,9 @@ const Product: FC = (): JSX.Element => {
     product?.title as string
   );
 
-  console.log('%croduct:::>', 'color:purple', product)
-
-  const guestSchema = yup.object().shape({
-    name: yup
-      .string()
-      .required("El nombre es requerido")
-      .min(4, "El nombre debe tener mínimo 4 caracteres"),
-    fatherSurname: yup
-      .string()
-      .required("El apellido paterno es requerido")
-      .min(4, "El apellido debe tener mínimo 4 caracteres"),
-    comment: yup
-      .string()
-      .required("El comentario es requerido")
-      .min(6, "El comentario debe tener mínimo 6 caracteres"),
-  });
-
-  const userSchema = yup.object().shape({
-    comment: yup
-      .string()
-      .required("El comentario es requerido")
-      .min(6, "El comentario debe tener mínimo 6 caracteres"),
-  });
+  useEffect(() => {
+    console.log('%croduct:::>', 'color:purple', product)
+  }, [product])
 
   const {
     handleSubmit,
@@ -171,6 +132,80 @@ const Product: FC = (): JSX.Element => {
                   stock={product?.stock as number}
                 />
               </Suspense>
+
+              <Accordion marginTop={'10px'} allowToggle >
+                <AccordionItem>
+                  <h2>
+                    <AccordionButton>
+                      <Box fontWeight={700} as="span" flex='1' textAlign='center'>
+                        Caracteristicas adicionales
+                      </Box>
+                      {/* <AddIcon fontSize='12px' /> */}
+                      <AccordionIcon />
+                    </AccordionButton>
+                  </h2>
+                  <AccordionPanel>
+                    {Object.keys(product ?? {}).map((element, index) => {
+                      const nameObject = `unidad-${element}`;
+                      const unidad = product[nameObject];
+                      const contentProduct = product[element];
+                      const sectionsExclude = ['category', 'description', 'price', 'tags', 'stock', 'title', 'image', 'subCategory']
+
+                      const returnDefault = (title: string, content: string) => {
+                        return (
+                          <Box display={'flex'} sx={{ alignItems: 'flex-end' }} key={index}>
+                            <Text fontWeight={700} fontSize='lg' mt={4}>{title + ": "}</Text>
+                            <Text fontWeight={500}>{content}</Text>
+                          </Box>
+                        )
+                      }
+
+                      if (element === 'Graduacion') {
+                        const longitudGrupo = Math.ceil((contentProduct?.length ?? 3) / 3);
+                        const newData = [contentProduct.substring(0, longitudGrupo), contentProduct.substring(longitudGrupo, longitudGrupo * 2), contentProduct.substring(longitudGrupo * 2)]
+                        return (
+                          <Box>
+                            <Text fontWeight={700} fontSize='lg' mt={4}>{element}</Text>
+                            {newData.map((e) => (
+                              <Badge sx={{ marginRight: '20px' }} variant='subtle' colorScheme='green' key={index * 3} >
+                                {e}
+                              </Badge>
+                            ))
+                            }
+                          </Box>
+                        )
+
+                      }
+                      if (element.includes('unidad') || contentProduct?.length === 0 || element === 'id' || sectionsExclude.includes(element)) return null
+                      if (Array.isArray(contentProduct)) {
+                        return (
+                          <Box>
+                            <Text fontWeight={700} fontSize='lg' mt={4}>{element}</Text>
+                            {
+                              contentProduct.map((e: [], i: any) => (
+                                <Badge sx={{ marginRight: '20px' }} variant='outline' colorScheme='green' key={i} >
+                                  {e}
+                                </Badge>
+                              ))
+                            }
+                          </Box>
+                        )
+                      }
+                      if (product[nameObject] !== undefined) {
+                        return returnDefault(element, (contentProduct + " " + unidad))
+                      }
+                      return (
+                        <Box key={index + 1}>
+                          {returnDefault(element, contentProduct)}
+                        </Box>
+                      )
+                    })
+                    }
+                  </AccordionPanel>
+                </AccordionItem>
+              </Accordion>
+
+
               <Box bgColor='white' p={6} borderRadius='lg' boxShadow='base'>
                 <Heading fontSize={[20, 24, 32]} mb={4}>
                   Comentarios sobre el producto
